Index structure & Vue object modified
The index.html page has been modified around the select and options so when a user selects a pronoun or accord, the data object inside the Vue instance is going to update its value.

The data object has been modified (accords were changed)  and added two methods.

Updated the readme so people will knwo what this project is about.