App should work in a more ethical way
Index.html : the login button now should work and launch the logMeInWithPermissions method and another button appeared to call the publishToFB method to publish content to FB. If it's working, it's way closer from the final result than the prev version.
Script.js :  not calling logMeInWithPermissions & publishToFB methods which was unethical (but it was for test purpose only !). Changed the "message" parameter that was given when calling FB.api to POST.