Changed the 'share' button behaviour and minor changes
Index.html : added a message that'll call the user to select a pronoun or an accord, removed the useless repetition of the app's name. Changed the 'share' button behaviour, it'll not appear until you have chosen both a pronoun and an accord. The message to preview what you'll post to FB won't appear until you did select a pronoun or an accord.
Style.scss : choices_previewer id is now a class and is called accented_text which is more coherent.
Script.js : removed the subtitle, added call_to_action_first that holds the first message pushing the user to do something with the app and slightly changed the message element.