Added a protection when sharing to FB and minor changes
Index.html : added a message that should tell the user if the status was submitted with success or not.
Script.js : added an element to the Vue object and modified the publishToFB method to reload the page in case of incorrect datas submision trial.
Style.css : added a new class with a color variant.